+++ src/ConfigMenu.cs
@@ -15,6 +15,7 @@ internal static class GGVConfig
   internal static bool FIX_DAMAGE_CAPS   = true;
   internal static bool FIX_EVOLVER       = true;
   internal static bool FIX_AMMO_DRIFT    = true;
+  internal static bool FIX_REPAUSE       = true;
 
   // Safe Optimizations
   internal static bool OPT_PROJ_STATUS   = true;
@@ -57,6 +58,7 @@ internal static class GGVConfig
     FIX_DAMAGE_CAPS   = "Enabled" == ConfigMenu._Gunfig.Value(ConfigMenu.DAMAGE_CAPS);
     FIX_EVOLVER       = "Enabled" == ConfigMenu._Gunfig.Value(ConfigMenu.EVOLVER);
     FIX_AMMO_DRIFT    = "Enabled" == ConfigMenu._Gunfig.Value(ConfigMenu.AMMO_DRIFT);
+    FIX_REPAUSE       = "Enabled" == ConfigMenu._Gunfig.Value(ConfigMenu.REPAUSE);
 
     OPT_VIS_CHECKS    = "Enabled" == ConfigMenu._Gunfig.Value(ConfigMenu.VIS_CHECKS);
     OPT_OCCLUSION     = "Enabled" == ConfigMenu._Gunfig.Value(ConfigMenu.OCCLUSION);
@@ -94,6 +96,7 @@ internal static class GGVConfig
     WriteLine($"FIX_DAMAGE_CAPS          = {FIX_DAMAGE_CAPS}");
     WriteLine($"FIX_EVOLVER              = {FIX_EVOLVER}");
     WriteLine($"FIX_AMMO_DRIFT           = {FIX_AMMO_DRIFT}");
+    WriteLine($"FIX_REPAUSE              = {FIX_REPAUSE}");
 
     WriteLine($"OPT_OCCLUSION            = {OPT_OCCLUSION}");
     WriteLine($"OPT_AMMO_DISPLAY         = {OPT_AMMO_DISPLAY}");
@@ -149,6 +152,7 @@ internal static class ConfigMenu
   internal const string DAMAGE_CAPS   = "Beam Damage Cap Fix";
   internal const string EVOLVER       = "Evolver Devolve Fix";
   internal const string AMMO_DRIFT    = "Ammo Drift Fix";
+  internal const string REPAUSE       = "Unpause / Repause Fix";
 
   internal const string SAFE_OPT      = "Safe Optimizations";
   internal const string OCCLUSION     = "Optimize Occlusion";
@@ -195,6 +199,7 @@ internal static class ConfigMenu
     sf.FancyToggle(DAMAGE_CAPS, "Fixes beams not ignoring boss damage caps\neven when set to do so. (No such\nbeam exists in vanilla, mostly for modded use).");
     sf.FancyToggle(EVOLVER, "Fixes Evolver devolving to its 2nd form\nafter dropping it, picking it back up,\nand killing 5 enemies to level it up.");
     sf.FancyToggle(AMMO_DRIFT, "Fixes ammo display drifting to the right when\na gun temporarily gets infinite ammo\n(e.g., from Magazine Rack).");
+    sf.FancyToggle(REPAUSE, "Fixes game continuing to run if you\nunpause and quickly repause during menu\nfading animation.");
 
     Gunfig so = _Gunfig.AddSubMenu(SAFE_OPT);
     so.FancyToggle(OCCLUSION, "Speeds up occlusion calculations by\nusing optimized algorithms and caching.\nSaves a large amount of CPU.");