+++ changelog.md
@@ -5,6 +5,7 @@
 - Added logging showing which GGV configuration options are enabled on startup
 - Further improved "Optimize Ammo Display" option
 - Added "Optimize Linear Cast" option, mitigating lag caused by having lots of very fast moving projectiles on the screen
+- Added "Unpause / Repause Fix"
 
 ## 1.3.3 (2025-06-01)